refactor: modularized test_connectivity and webhooks and added helper to get aurthorization code flow
